==> 00_guide to analysis.R <==
# ########################################################### #
########Prep for final methods and broad-scale analysis########
# ########################################################### #

# ############################ #
########Data preparation########
# ############################ #

#######Data integration#######
#What metrics are available for each region?
#What could or should be added? (e.g., GIS)
#Work through when and how do we corsswalk indicators that are measured differently?
###AW, WM GP
###NESE
###PNW
#Add drainage area

##Roll up non-PNW SDAM data for PNW resolution
#BMI
#Algae and streamer mosses

#######Data review and checking#######
#Check coordinate changes between visits
#Make new site codes if necessary
#For GP, NESE???


#######GP/NESE data cleanup recheck to ensure (Ken)#######
#Logger data exports are complete
#Logger versus sampling dates line up
#Logger data complete enough for metric calculation
#Calibration completeness

#######Parsing the sites by different boundaries for performance testing and potential stratification#######
#Explore these regions
##MLRA
##NCCA
##OHWM
##Snow

#######Data completeness
